JBPM-5819: Case Modeller: OOME: Cannot create ImageData when dragging shapes
ProjectDiagramExplorerScreen has a TreeExplorer widget that includes a Node's Glyph.
Some glyphs use HTML images that are added to the DOM root and copied from the DOM
to the Canvas. Each instance of each glyph has a new DOM element created. TreeExplorer
creates more and more TreeExplorerItems containing Glyphs as the User interacts with
the editor. Furthermore memory used by SVG images is not released by the browser

This PR addresses release of Glyphs by the TreeExplorer and replaces CM's use of SVGs
with a single PNG. Memory used by non-SVG images is correctly released by the browser.