JBPM-5828: Case Modeller: Stages and Activities to use icons
Whilst adding icons to CM's shapes it was noticed that DOM
elements created representing Glyph Pictures for the Palette
and Shapes children were not released when the editor closed.
Thus when a file, or multiple files, were opened and closed
a growing number of unreleased DOM elements remained.

This PR includes fixes to release DOM elements and may, in
part, relate to https://issues.jboss.org/browse/JBPM-5462