RHPAM-3544 - Data Object editor removes imports when saving
It is possible to have an element loaded on the properties tab and not selected on the canvas.
The assignments widget attempts to load data objects based on that canvas item. If there is none, the error happens.
This fix consists in refreshing the properties tab correctly and displaying the selected node properly.